api/requirements.txt
fastapi
requests
ximilar-client
uvicorn[standard]

